feed and activity added

+++ base/views.py
@@ -1,41 +1,64 @@
 from django.shortcuts import render, redirect
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 
 
-
-def home(request):
+def home(request): 
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
-        Q(description__icontains=q) |
-        Q(host__username__icontains=q)
+        Q(description__icontains=q)
         )
+    room_messages = Message.objects.filter(
+        Q(room__topic__name__icontains=q) |
+        Q(room__name__icontains=q)
+    )
     topics = Topic.objects.all()
     room_count  = rooms.count()
-    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count, 'room_messages': room_messages}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {'room': room}
+    comments = room.message_set.all()
+    participants = room.participants.all()
+    
+    if request.method == 'POST':
+        new_message = Message.objects.create(
+            user =  request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        new_message.save()
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
+    context = {'room': room, 'comments': comments, 'participants': participants}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
         form = RoomForm(request.POST)
         if form.is_valid():
-            form.save()
+            room = form.save(commit=False)
+            room.host = request.user
+            room.save()
             return redirect('home')
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if room.host != request.user:
+        return HttpResponse("You can't perform this action")
     form = RoomForm(instance=room)
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -44,11 +67,33 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if room.host != request.user:
+        return HttpResponse("You can't perform this action")
     context = {'obj': room}
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', context)
+
+@login_required(login_url='/')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user  != message.user:
+        return HttpResponse("Hey, You can't do that!")
+    room = message.room.id
+    if request.method == 'POST':
+        message.delete()
+        messages.success(request, "Message deleted!")
+        return redirect('room', room)
+    return render(request, 'base/delete.html', {'obj': message})
+
+def userprofile(request, pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.message_set.all()
+    topics = Topic.objects.all()
+    context = {'user': user, 'rooms': rooms, 'room_messages': room_messages, 'topics': topics}
+    return render(request, 'users/profile.html', context)
